+++ README.md
@@ -1,12 +1,33 @@
-# 문자열 덧셈 계산기
+# 키친포스
 
-## 요구사항
-[O] 쉼표(,) 또는 콜론(:)을 구분자로 가지는 문자열을 전달하는 경우 구분자를 기준으로 분리한 각 숫자의 합을 반환 (예: “” => 0, "1,2" => 3, "1,2,3" => 6, “1,2:3” => 6)
-[O] 앞의 기본 구분자(쉼표, 콜론)외에 커스텀 구분자를 지정할 수 있다. 커스텀 구분자는 문자열 앞부분의 “//”와 “\n” 사이에 위치하는 문자를 커스텀 구분자로 사용한다. 예를 들어 “//;\n1;2;3”과 같이 값을 입력할 경우 커스텀 구분자는 세미콜론(;)이며, 결과 값은 6이 반환되어야 한다.
-[O] 문자열 계산기에 숫자 이외의 값 또는 음수를 전달하는 경우 RuntimeException 예외를 throw한다.
-[O] 빈 문자열 또는 null 값을 입력할 경우 0을 반환해야 한다.(예 : “” => 0, null => 0)
-[O] 숫자 하나를 문자열로 입력할 경우 해당 숫자를 반환한다.(예 : “1”)
-[O] 숫자 두개를 컴마(,) 구분자로 입력할 경우 두 숫자의 합을 반환한다.(예 : “1,2”)
-[O] 구분자를 컴마(,) 이외에 콜론(:)을 사용할 수 있다. (예 : “1,2:3” => 6)
-[O]  “//”와 “\n” 문자 사이에 커스텀 구분자를 지정할 수 있다. (예 : “//;\n1;2;3” => 6)
-[O] 음수를 전달할 경우 RuntimeException 예외가 발생해야 한다. (예 : “-1,2,3”)
+## 퀵 스타트
+
+```sh
+cd docker
+docker compose -p kitchenpos up -d
+```
+
+## 요구 사항
+- 식당 음식 판매 정보 관리 기기 구현하기
+- 메뉴 그룹
+  - 메뉴 그룹이란 메뉴들의 집합을 의미한다.
+  - 사용자는 새로운 메뉴 그룹을 추가할 수 있다.
+  - 사용자는 모든 메뉴 그룹을 조회할 수 있다.
+- 메뉴
+  - 메뉴란 식당에서 판매되고 있는 음식과 해당 음식과 관련된 정보들이다.
+  - 사용자는 새로운 메뉴를 추가할 수 있다.
+  - 사용자는 메뉴의 가격을 변경할 수 있다.
+  - 사용자는 메뉴를 숨겨 선택적으로 보여줄 수 있다.
+- 주문
+  - 주문이란 식당에 있는 메뉴에 돈을 지불하여 음식을 주문한 것을 의미한다.
+  - 사용자는 식당에서 먹는 것과 포장 그리고 배달에 따라 선택적으로 주문할 수 있다.
+  - 주문을 하게 되면 주문 대기 상태에서 주문이 가능하다면 접수 상태로 변경이 된다.
+  - 
+
+## 용어 사전
+
+| 한글명 | 영문명 | 설명 |
+| --- | --- | --- |
+|  |  |  |
+
+## 모델링
